Update port 8080 remap for Lithium
Lithium changed how ODL opens port 8080, so the config needed to be
updated to match.

This is needed to support running ODL with other apps that rely on
8080.

Ideally, we'll eventually support remapping 8181 as well.

We also need to clarify our API for doing port remapping,
as the var name is horrible.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>