Refactor logic for running in containers
New CentOS containers (7.2.1511 and not 7.1.1503) ship full systemd,
instead of using fakesystemd or not including it. It doesn't work
without modifying the Dockerfile and/or passing special params to the
`docker run` command, and it's generally not recommended for use in
containers, so we don't want to use it when provisioning a container.

The old checks were seeing systemd and (rightly so, it's installed) and
attempting to use it in containers. The new one directly checks if
running in a container instead of inferring from presence or absence of
systemd.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>