Don't use template for Karaf feature config
Templates don't work well now that ansible-opendaylight supports
installing multiple versions of ODL (via rpm_repo*, for example).

The content of the tempalted file is completly replaced, when only the
exact line that needs to be changed should be managed and the rest of
the file should use the version's default content.

Changing the version of ODL installed via params also requires changing
the default Karaf feature repos to match the version. This is confusing
and unnecessary - no one uses this knob that I know of.

Signed-off-by: Daniel Farrell <dfarrell@redhat.com>